+++ src/com/lld/delayedscheduler/queue/CustomQueue.java
@@ -1,4 +1,4 @@
-package com.lld.delayedscheduler;
+package com.lld.delayedscheduler.queue;
 
 public interface CustomQueue<E extends Comparable<E>> {
     void add(E e) throws InterruptedException;
