+++ src/com/lld/delayedscheduler/queue/PriorityQueueBasedBlockingQueue2.java
@@ -1,4 +1,4 @@
-package com.lld.delayedscheduler;
+package com.lld.delayedscheduler.queue;
 
 import java.util.PriorityQueue;
 import java.util.concurrent.locks.Condition;
