+++ src/com/lld/delayedscheduler/scheduler/DelayedScheduler.java
@@ -1,4 +1,4 @@
-package com.lld.delayedscheduler;
+package com.lld.delayedscheduler.scheduler;
 
 public interface DelayedScheduler {
     void schedule(Runnable task, long delayMillis) throws InterruptedException;
