+++ src/com/lld/delayedscheduler/scheduler/PriorityQueueBasedDelayedScheduler.java
@@ -1,21 +1,28 @@
-package com.lld.delayedscheduler;
+package com.lld.delayedscheduler.scheduler;
 
-import java.util.concurrent.*;
 
-public class DelayedSchedulerImpl implements DelayedScheduler {
-    private DelayQueue<Delayed> queue;
+import com.lld.delayedscheduler.queue.CustomQueue;
+import com.lld.delayedscheduler.queue.PriorityQueueBasedBlockingQueue;
+
+import java.util.concurrent.Delayed;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.TimeUnit;
+
+public class PriorityQueueBasedDelayedScheduler implements DelayedScheduler {
+    private CustomQueue<Delayed> queue;
     private ExecutorService executorService;
     private SharedVariable sharedVariable;
 
-    public DelayedSchedulerImpl() {
-        this.queue = new DelayQueue<>();
+    public PriorityQueueBasedDelayedScheduler() {
+        this.queue = new PriorityQueueBasedBlockingQueue<>();
         this.executorService = Executors.newFixedThreadPool(10);
         this.sharedVariable = new SharedVariable();
-        this.executorService.submit(new TaskRunner(this.queue, this.executorService, this.sharedVariable));
+        this.executorService.submit(new TaskRunner2(this.queue, this.executorService, this.sharedVariable));
     }
 
     @Override
-    public void schedule(Runnable task, long delayMillis) {
+    public void schedule(Runnable task, long delayMillis) throws InterruptedException {
         DelayedTask delayedTask = new DelayedTask(task, delayMillis);
         this.queue.add(delayedTask);
     }
@@ -35,10 +42,10 @@ public class DelayedSchedulerImpl implements DelayedScheduler {
             // to forcefully interrupt all the current threads.
             // If currently running threads throw `RunTimeException` once interrupted,
             // the system will shut down completely else it will still keep on running
-            if (!this.executorService.awaitTermination(1, TimeUnit.MINUTES)) {
+            if (!this.executorService.awaitTermination(30, TimeUnit.SECONDS)) {
                 this.executorService.shutdownNow();
 
-                if (!this.executorService.awaitTermination(1, TimeUnit.MINUTES)) {
+                if (!this.executorService.awaitTermination(30, TimeUnit.SECONDS)) {
                     System.out.println("Not stopped");
                 }
             }
@@ -49,17 +56,17 @@ public class DelayedSchedulerImpl implements DelayedScheduler {
     }
 
     @Override
-    public int pendingTasks() {
+    public synchronized int pendingTasks() {
         return this.queue.size();
     }
 }
 
-class TaskRunner implements Runnable {
-    private DelayQueue<Delayed> queue;
+class TaskRunner2 implements Runnable {
+    private CustomQueue<Delayed> queue;
     private ExecutorService executorService;
     private SharedVariable sharedVariable;
 
-    public TaskRunner(DelayQueue<Delayed> queue, ExecutorService executorService, SharedVariable sharedVariable) {
+    public TaskRunner2(CustomQueue<Delayed> queue, ExecutorService executorService, SharedVariable sharedVariable) {
         this.queue = queue;
         this.executorService = executorService;
         this.sharedVariable = sharedVariable;
@@ -68,10 +75,16 @@ class TaskRunner implements Runnable {
     @Override
     public void run() {
         while (!this.sharedVariable.isStop()) {
-            DelayedTask task = (DelayedTask) this.queue.poll();
+            DelayedTask delayedTask;
 
-            if (task != null) {
-                this.executorService.submit(task.getTask());
+            try {
+                delayedTask = (DelayedTask) this.queue.poll();
+                if (delayedTask != null) {
+                    this.executorService.submit(delayedTask.getTask());
+                }
+            } catch (InterruptedException e) {
+                Thread.currentThread().interrupt();
+                throw new RuntimeException(e);
             }
         }
     }
